Gitignored the file that I just nuked.
(ie. „Spectre” or „universal-installer-scaffolding.sh”)

Why would I ignore a file, if it's not here, anyway? Well... I linked it back to /scripts/universal-installer-scaffolding.sh from the newly-created repo made just for it. But it's not SUPPOSED TO be a part of this repo, so I had to gitignore the said link.

Why would I link it, if it isn't suppoed to be here, anyway? It's for ease of development, mostly. Spectre IS a completly separate project, that can be used entirely stand-alone from this one and vice-versa (so they have different repos), but the are supposed to be developed together, and constantly switching vscode windows would just get annoying.